feat(frontend): enhance PublishAgentModal with user authentication prompts
- Integrated user authentication checks in the PublishAgentModal component.
- Added loading and authentication prompts to guide users in logging in or creating an account.
- Removed redundant user checks from the AgentSelectStep component to streamline the user experience.
